Remove legacy openapi specs Organize imports from right generated files
