--- .github/workflows/pr_opex_api_backend.yml
@@ -1,37 +1,0 @@
-name: PR - Opex Dashboard api_backend
-
-on:
-  pull_request:
-    types:
-      - opened
-      - edited
-      - synchronize
-      - reopened
-    branches:
-      - master
-    paths:
-      - 'api_backend.yaml'
-      - '.github/workflows/pr_opex_api_backend.yml'
-      - '.github/workflows/opex_api_backend.yml'
-      - '.opex/api_backend/**'
-
-  workflow_dispatch:
-
-concurrency:
-  group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}
-  cancel-in-progress: false
-
-permissions:
-  id-token: write
-  pull-requests: write
-
-jobs:
-  plan:
-    uses: ./.github/workflows/call_pr_opex_api.yml
-    name: Plan changes against Production
-    secrets:
-      AZURE_CLIENT_ID_CI: ${{ secrets.AZURE_CLIENT_ID_CI }}
-      AZURE_TENANT_ID: ${{ secrets.AZURE_TENANT_ID }}
-      AZURE_SUBSCRIPTION_ID: ${{ secrets.AZURE_SUBSCRIPTION_ID }}
-    with:
-      api_name: api_backend
