--- api_parameters.yaml
@@ -1,66 +1,0 @@
-swagger: "2.0"
-info:
-  title: "IO API parameters"
-  version: "0.1"
-paths:
-  "/dummy":
-    get:
-      responses:
-        "200":
-          description: "ok"
-definitions:
-  GetMessagesParameters:
-    type: object
-    title: GetMessagesParameters
-    description: Describes the GetMessages api query parameters
-    properties:
-      pageSize:
-        $ref: "#/definitions/NonNegativeIntegerFromString"
-      enrichResultData:
-        $ref: "#/definitions/BooleanFromString"
-      getArchivedMessages:
-        $ref: "#/definitions/BooleanFromString"
-      maximumId:
-        $ref: "#/definitions/Ulid"
-      minimumId:
-        $ref: "#/definitions/Ulid"
-  GetMessageParameters:
-    type: object
-    title: GetMessageParameters
-    description: Describes the GetMessage api parameters
-    properties:
-      id:
-        $ref: "#/definitions/Ulid"
-      public_message:
-        $ref: "#/definitions/BooleanFromString"
-    required:
-      - id
-  GetPublishedCategoriesParameters:
-    type: object
-    title: GetPublishedCategoriesParameters
-    description: Describes the GetMessages api query parameters
-    properties:
-      count_new_discounts:
-        $ref: "#/definitions/BooleanFromString"
-  NonNegativeIntegerFromString:
-    type: string
-    description: Represents a non negative integer parsed from a string
-    format: NonNegativeIntegerFromString
-    x-import: '@pagopa/ts-commons/lib/numbers'
-    example: 10
-  BooleanFromString:
-    type: string
-    description: Represents a boolean parsed from a string
-    format: BooleanFromString
-    # TODO: Import from ts-commons when it will be updated during the
-    # update of backend to fp-ts 2.x
-    x-import: '../../src/types/booleans'
-    example: true
-  NonEmptyString:
-    type: string
-    description: Represents a non empty string
-    format: NonEmptyString
-    x-import: '@pagopa/ts-commons/lib/strings'
-    example: "xxxxxxxx"
-  Ulid:
-    $ref: "https://raw.githubusercontent.com/pagopa/io-functions-commons/v28.6.0/openapi/definitions.yaml#/Ulid"
