--- openapi/api_notifications.template.yaml
@@ -1,83 +1,0 @@
-swagger: "2.0"
-info:
-  version: 16.7.0
-  title: Proxy API
-  description: Mobile and web proxy API gateway.
-host: localhost
-basePath: /api/v1
-schemes:
-  - https
-security:
-  - Token: []
-paths:
-  "/dummy":
-    get:
-      responses:
-        "200":
-          description: "ok"
-definitions:
-  CreatedMessageWithContent:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/CreatedMessageWithContent"
-  CreatedMessageWithoutContent:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/CreatedMessageWithoutContent"
-  TimeToLiveSeconds:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/TimeToLiveSeconds"
-  SenderMetadata:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/SenderMetadata"
-  ServiceName:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ServiceName"
-  OrganizationName:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/OrganizationName"
-  DepartmentName:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/DepartmentName"
-  Timestamp:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/Timestamp"
-  FiscalCode:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/FiscalCode"
-  MessageContent:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/MessageContent"
-  MessageSubject:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/MessageSubject"
-  MessageBodyMarkdown:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/MessageBodyMarkdown"
-  PaymentNoticeNumber:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/PaymentNoticeNumber"
-  PaymentAmount:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/PaymentAmount"
-  PaymentData:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/PaymentData"
-  ProblemJson:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ProblemJson"
-  NotificationMessage:
-    x-one-of: true
-    allOf:
-      - $ref: '#/definitions/CreatedMessageWithContent'
-      - $ref: '#/definitions/CreatedMessageWithoutContent'
-  Notification:
-    title: Notification
-    description: A received Notification.
-    type: object
-    properties:
-      message:
-        $ref: '#/definitions/NotificationMessage'
-      sender_metadata:
-        $ref: '#/definitions/SenderMetadata'
-    required:
-      - message
-      - sender_metadata
-  SuccessResponse:
-    type: object
-    properties:
-      message:
-        type: string
-responses: {}
-parameters: {}
-consumes:
-  - application/json
-produces:
-  - application/json
-securityDefinitions:
-  Token:
-    type: apiKey
-    name: token
-    in: query
