--- openapi/api_pn.template.yaml
@@ -1,83 +1,0 @@
-swagger: "2.0"
-info:
-  version: 16.7.0
-  title: Piattaforma Notifiche
-host: localhost
-basePath: /api/v1/pn
-schemes:
-  - https
-security:
-  - Bearer: []
-paths:
-  "/activation":
-    get:
-      operationId: getPNActivation
-      summary: Get Piattaforma Notifiche Activation
-      description: |
-        Get the Piattaforma Notifiche activation for Special Service "Avvisi di Cortesia"
-      parameters:
-        - in: query
-          name: isTest
-          required: false
-          type: boolean
-      responses:
-        "200":
-          description: Request created.
-          schema:
-            $ref: "#/definitions/PNActivation"
-        "401":
-          description: Bearer token null or expired.
-        "429":
-          description: Too many requests
-        "500":
-          description: Internal Server Error
-          schema:
-            $ref: "#/definitions/ProblemJson"
-    post:
-      operationId: upsertPNActivation
-      summary: Upsert Piattaforma Notifiche Activation
-      description: |
-        Upsert the Piattaforma Notifiche activation for Special Service "Avvisi di Cortesia"
-      parameters:
-        - in: body
-          name: body
-          required: true
-          schema:
-            $ref: '#/definitions/PNActivation'
-          x-examples:
-            application/json:
-              activation_status: true
-        - in: query
-          name: isTest
-          required: false
-          type: boolean
-      responses:
-        "204":
-          description: Request created.
-        "400":
-          description: Bad request
-          schema:
-            $ref: "#/definitions/ProblemJson"
-        "401":
-          description: Bearer token null or expired.
-        "429":
-          description: Too many requests
-        "500":
-          description: Internal Server Error
-          schema:
-            $ref: "#/definitions/ProblemJson"
-definitions:
-  ProblemJson:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ProblemJson"
-  PNActivation:
-    type: object
-    properties:
-      activation_status:
-        type: boolean
-    required:
-      - activation_status
-securityDefinitions:
-  Bearer:
-    type: apiKey
-    name: Authorization
-    in: header
