+++ package.json
@@ -28,13 +28,11 @@
     "openapi:bundle": "./openapi/bundle.sh $npm_package_version",
     "generate": "npm-run-all generate:proxy:* generate:api:* generate:messages:* generate:lollipop-definitions generate:lollipop-first-sign",
     "generate:dev-certs": "sh ./scripts/generate-test-certs.sh certs",
-    "generate:proxy:api-models": "rimraf generated/backend && gen-api-models --api-spec api_backend.yaml --out-dir generated/backend",
+    "generate:proxy:api-models": "rimraf generated/backend && gen-api-models --api-spec openapi/generated/api_backend.yaml --out-dir generated/backend",
     "generate:proxy:identity-api-models": "rimraf generated/identity && gen-api-models --api-spec openapi/generated/api_identity.yaml --out-dir generated/identity",
     "generate:proxy:communication-api-models": "rimraf generated/communication && gen-api-models --api-spec openapi/generated/api_communication.yaml --out-dir generated/communication",
-    "generate:proxy:api-parameters": "gen-api-models --api-spec api_parameters.yaml --out-dir generated/parameters",
-    "generate:proxy:notification-models": "rimraf generated/notifications && gen-api-models --api-spec api_notifications.yaml --out-dir generated/notifications",
+    "generate:proxy:api-parameters": "gen-api-models --api-spec openapi/generated/api_parameters.yaml --out-dir generated/parameters",
     "generate:messages:notifications": "rimraf generated/messages && gen-api-models --api-spec notification_queue_messages.yaml --out-dir generated/messages",
-    "generate:proxy:first-lc-models": "rimraf generated/first-lc-proxy-models && gen-api-models --api-spec api_lollipop_first_consumer.yaml --out-dir generated/first-lc-proxy-models",
     "generate:proxy:public-models": "rimraf generated/public  && gen-api-models --api-spec openapi/generated/api_platform_legacy.yaml --out-dir generated/public",
     "generate:api:io-messages": "rimraf generated/io-messages-api && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-functions-app-messages/v5.3.0/openapi/index.yaml --no-strict --out-dir generated/io-messages-api --request-types --response-decoders --client",
     "generate:proxy:cgn-card-platform-api-models": "rimraf generated/cgn-card-platform && gen-api-models --api-spec openapi/generated/api_cgn_card_platform.yaml --out-dir generated/cgn-card-platform",
@@ -52,7 +50,6 @@
     "generate:proxy:third-party-service": "rimraf generated/third-party-service && gen-api-models --api-spec openapi/consumed/api-third-party.yaml --out-dir generated/third-party-service --request-types --response-decoders --client",
     "generate:proxy:remote-content-service": "rimraf generated/remote-content-service && gen-api-models --api-spec openapi/consumed/api_remote_content.yaml --out-dir generated/remote-content-service --request-types --response-decoders --client",
     "generate:proxy:piattaforma-notifiche": "rimraf generated/piattaforma-notifiche && gen-api-models --api-spec openapi/consumed/api-piattaforma-notifiche.yaml --out-dir generated/piattaforma-notifiche --request-types --response-decoders --client",
-    "generate:api:piattaforma-notifiche-courtesy": "rimraf generated/api_piattaforma-notifiche-courtesy && gen-api-models --api-spec api_pn.yaml --out-dir generated/api_piattaforma-notifiche-courtesy",
     "generate:proxy:piattaforma-notifiche-courtesy": "rimraf generated/piattaforma-notifiche-courtesy && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/pn-user-attributes/13b91a353f6a219c60d5cdadc023cc18c4157554/docs/openapi/pn-address-book-b2b-external-io-v1.yaml --out-dir generated/piattaforma-notifiche-courtesy --request-types --response-decoders --client",
     "generate:test-certs": "sh ./scripts/generate-test-certs.sh certs",
     "generate:lollipop-definitions": "rimraf generated/lollipop && gen-api-models --api-spec openapi/lollipop_definitions.yaml --out-dir generated/lollipop",
