+++ src/controllers/__tests__/pnController.test.ts
@@ -1,7 +1,7 @@
 import { IResponseType } from "@pagopa/ts-commons/lib/requests";
 import { FiscalCode } from "@pagopa/ts-commons/lib/strings";
 import { right, left } from "fp-ts/lib/Either";
-import { PNActivation } from "../../../generated/api_piattaforma-notifiche-courtesy/PNActivation";
+import { SendActivation } from "../../../generated/communication/SendActivation";
 import { IoCourtesyDigitalAddressActivation } from "../../../generated/piattaforma-notifiche-courtesy/IoCourtesyDigitalAddressActivation";
 import { PNEnvironment } from "../../clients/pn-clients";
 import mockReq from "../../__mocks__/request";
@@ -118,7 +118,7 @@ describe("pnController#upsertPNActivationController", () => {
       const controller = upsertPNActivationController(mockUpsertPnActivation);
       req.body = {
         activation_status: true
-      } as PNActivation;
+      } as SendActivation;
       req.query = {
         isTest: isTestValue
       };
