+++ src/controllers/__tests__/profileController.test.ts
@@ -7,7 +7,7 @@ import {
 import * as E from "fp-ts/lib/Either";
 import { mockedUser } from "../../__mocks__/user_mock";
 import { EmailAddress } from "../../../generated/identity/EmailAddress";
-import { ExtendedProfile } from "../../../generated/backend/ExtendedProfile";
+import { ExtendedProfile } from "../../../generated/identity/ExtendedProfile";
 import { FiscalCode } from "../../../generated/identity/FiscalCode";
 import { IsInboxEnabled } from "../../../generated/identity/IsInboxEnabled";
 import { IsWebhookEnabled } from "../../../generated/identity/IsWebhookEnabled";
